+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 - "You miss 100 percent of the chances you don't take. — Wayne Gretzky" — Michael Scott
 - [react] feat: Export `createReduxEnhancer` to log redux actions as breadcrumbs, and attach state as an extra. (#2717)
 - [tracing] feat: `Add @sentry/tracing` (#2719)
-- [gatsby] fix: Make APM optional in gatsby package
-- [tracing] ref: Use idleTimout if no activities occur in idle transaction
+- [gatsby] fix: Make APM optional in gatsby package (#2752)
+- [tracing] ref: Use idleTimout if no activities occur in idle transaction (#2752)
 
 ## 5.19.2